apps/api/pyproject.toml
[project]
name = "vandyguessr-backend"
version = "0.1.0"
description = "VandyGuessr API - A GeoGuessr-style game for Vanderbilt University"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "motor>=3.4.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.2.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "boto3>=1.34.0",
    "pillow>=10.3.0",
    "pillow-heif>=0.16.0",
    "lagom>=2.7.7",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4.4",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
